Update Client::invoke_service to be more flexible.
Add ability to specify:

1. HTTP method
2. Content-Type
3. Queries

It is currently a very stringly-typed API but I am not sure if there's even an attempt to enumerate all the possible Content-Type.
Other than that, I am not sure if passing the queries directly as a String is correct here, because
we want to utilize stronger correctness with something like Iterator<Item=(String,String)> instead.
But this does give the freedom to the user to decide how they want to allocate the string.

Signed-off-by: Hanif Ariffin <hanif.ariffin.4326@gmail.com>